DWIM 82 - V2 - CONVERSION A SOFTWARE HOTELERO (NUEVAS TAREAS 10FEB

+++ app/Http/Controllers/BookController.php
@@ -626,22 +626,23 @@ class BookController extends AppController
           $oVisa = DB::table('book_visa')
                     ->where('book_id',$book->id)
                     ->first();
-            
             if ($oVisa){
               $hasVisa = true;
-              $visaData = json_decode($oVisa->visa_data);
-              if ($visaData){
-                foreach ($visaData as $k=>$v){
+              $visaData = json_decode($oVisa->visa_data,true);
+              $fieldsCard = ["name","number",'date',"cvc",'type'];
+              foreach ($fieldsCard as $f){
+                if (isset($visaData[$f])){
+                  if ($f == 'date') $visaData[$f] = str_replace ('/20', ' / ', $visaData[$f]);
                   $visaHtml .= '
-                  <div>
-                  <label>'.$k.'</label>
-                  <input type="text" class="form-control" value="'.$v.'" >
-                  <button class="btn btn-success copy_data" type="button">CP</button>
-                  </div>';
+                    <div>
+                    <label>'.$f.'</label>
+                    <input type="text" class="form-control" value="'.$visaData[$f].'" >
+                    <button class="btn btn-success copy_data" type="button"><i class="fa fa-copy"></i></button>
+                    </div>';
                 }
-               
               }
             }
+            
           }
           
        } else {
@@ -1294,7 +1295,13 @@ class BookController extends AppController
         try
         {
             $book = \App\Book::find($id);
-
+            if (count($book->pago) > 0){
+               return [
+                    'status'   => 'danger',
+                    'title'    => 'Error:',
+                    'response' => "La Reserva posee cargos asociados."
+                ];
+            }
             foreach ($book->notifications as $key => $notification)
             {
                 $notification->delete();
@@ -2656,6 +2663,7 @@ class BookController extends AppController
      */
     function getVisa(Request $request){
         $booking = $request->input('booking', null);
+        $fieldsCard = ["name","number",'date',"cvc",'type'];
         if ($booking){
           $aux = explode('-', $booking);
           if (is_array($aux) && count($aux) == 2){
@@ -2668,15 +2676,19 @@ class BookController extends AppController
                     ->first();
             
             if ($oVisa){
-              $visaData = json_decode($oVisa->visa_data);
+              $visaData = json_decode($oVisa->visa_data, true);
+              
               if ($visaData){
-                foreach ($visaData as $k=>$v){
-                  echo '
-                  <div>
-                  <label>'.$k.'</label>
-                  <input type="text" class="form-control" value="'.$v.'" >
-                  <button class="btn btn-success copy_data" type="button">CP</button>
-                  </div>';
+                foreach ($fieldsCard as $f){
+                  if (isset($visaData[$f])){
+                    if ($f == 'date') $visaData[$f] = str_replace ('/20', ' / ', $visaData[$f]);
+                    echo '
+                      <div>
+                      <label>'.$f.'</label>
+                      <input type="text" class="form-control" value="'.$visaData[$f].'" >
+                      <button class="btn btn-success copy_data" type="button"><i class="fa fa-copy"></i></button>
+                      </div>';
+                  }
                 }
                 return ;
               }
@@ -2693,24 +2705,28 @@ class BookController extends AppController
                 $creditCard = $oZodomus->reservations_cc($booking->propertyId,$booking->external_id);
 //                $creditCard = json_decode('{"status":{"returnCode":"200","returnMessage":"OK","channelLogId":"","channelOtherMessages":"","timestamp":"2020-02-03 08:36:04"},"reservation":{"id":"2493440995"},"customerCC":{"name":"","cvc":"","number":"","date":"","type":""}}');
                 if ($creditCard && isset($creditCard->status) && $creditCard->status->returnCode == 200){
-                  $visa_data = $creditCard->customerCC;
+                  $visa_data = json_encode($creditCard->customerCC);
                   DB::table('book_visa')->insert([
                     'book_id' =>$bookingID,
                     'user_id'=>Auth::user()->id,
                     'customer_id'=>$clientID,
-                    'visa_data'=>json_encode($visa_data),
+                    'visa_data'=>($visa_data),
                     'created_at'=>date('Y-m-d H:m:s'),
                     'updated_at'=>date('Y-m-d H:m:s'),
                    ]);
                   
                   if ($visa_data){
-                    foreach ($visa_data as $k=>$v){
-                      echo '
-                      <div>
-                      <label>'.$k.'</label>
-                      <input type="text" class="form-control" value="'.$v.'" >
-                      <button class="btn btn-success copy_data" type="button">CP</button>
-                      </div>';
+                    $visa_data = json_decode($visa_data, true);
+                    foreach ($fieldsCard as $f){
+                      if (isset($visaData[$f])){
+                        if ($f == 'date') $visaData[$f] = str_replace ('/20', ' / ', $visaData[$f]);
+                        echo '
+                          <div>
+                          <label>'.$f.'</label>
+                          <input type="text" class="form-control" value="'.$visaData[$f].'" >
+                          <button class="btn btn-success copy_data" type="button"><i class="fa fa-copy"></i></button>
+                          </div>';
+                      }
                     }
                   }
                 }
